Implementation a reducer that removes a book
