+++ src/redux/books/booksSlice.js
@@ -1,7 +1,26 @@
+import { v4 as uuidv4 } from 'uuid';
 import { createSlice } from '@reduxjs/toolkit';
 
+const books = [
+  {
+    id: uuidv4(),
+    title: 'The 100',
+    author: 'Alexis Walner',
+  },
+  {
+    id: uuidv4(),
+    title: 'Introduction to Algorithms',
+    author: 'Thomas H. Corme',
+  },
+  {
+    id: uuidv4(),
+    title: 'Computer Networks',
+    author: 'Andrew S. Tanenbaum',
+  },
+];
+
 const initialState = {
-  bookItems: [],
+  bookItems: books,
   isLoading: true,
 };
