+++ src/routes/BookPage.js
@@ -1,29 +1,14 @@
-import { v4 as uuidv4 } from 'uuid';
+import { useSelector } from 'react-redux';
 import Books from '../components/Books';
 import Input from '../components/Input';
 
 const BookPage = () => {
-  const books = [
-    {
-      id: uuidv4(),
-      title: 'The 100',
-      author: 'Alexis Walner',
-    },
-    {
-      id: uuidv4(),
-      title: 'Introduction to Algorithms',
-      author: 'Thomas H. Corme',
-    },
-    {
-      id: uuidv4(),
-      title: 'Computer Networks',
-      author: 'Andrew S. Tanenbaum',
-    },
-  ];
+  const { bookItems, isLoading } = useSelector((state) => state.book);
+  console.log(`*****: ${JSON.stringify(bookItems)}`);
 
   return (
     <section>
-      <Books books={books} />
+      <Books books={JSON.stringify(bookItems)} />
       <Input />
     </section>
   );
